refactor(rules/inline): simple markup inline rules
