+++ src/rules/inline/emphasis.ts
@@ -1,17 +1,6 @@
 import Renderer from 'markdown-it/lib/renderer';
-import Token from 'markdown-it/lib/token';
 
-const rules = new Set(['em_open', 'em_close', 'strong_open', 'strong_close']);
-
-function basic(tokens: Token[], i: number) {
-    const {type, markup} = tokens[i];
-
-    if (rules.has(type)) {
-        return markup;
-    }
-
-    throw new Error('failed to render emphasis');
-}
+import {basic} from 'src/rules/inline/basic';
 
 const emphasis: Renderer.RenderRuleRecord = {
     em_open: basic,
